backend/requirements.txt
Flask==3.0.3
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
Flask-Marshmallow==1.2.1
Marshmallow==3.21.2
marshmallow-sqlalchemy==1.4.2
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
python-dotenv==1.0.1
celery==5.4.0
redis==5.0.7
pytest==8.3.2
requests==2.32.3
